refactor getting status into its own function
